+++ vsc-base/out/vsc-base-raw.d.ts
@@ -334,21 +334,21 @@ export declare const getJSONCircularReplacer: () => (_key: string, value: unknow
  * @description
  * Stringify an object. \
  * Uses JSON.stringify and the circular ref safe replacer (see vsc.getJSONCircularReplacer)
- * @see http://vsc-base.org/#toString
+ * @see http://vsc-base.org/#toJSONString
  * @param obj
  * @param replacer
  * @param space
  * @vscType Raw
  * @debugTool Primary a debugging method.
- * @oneLineEx const objString = vsc.toString(someObject);
+ * @oneLineEx const objString = vsc.toJSONString(someObject);
  * @returns string
  */
-export declare const toString: (obj: any, replacer?: (_key: string, value: unknown) => unknown, space?: number, maxDepth?: number) => string;
+export declare const toJSONString: (obj: any, replacer?: (_key: string, value: unknown) => unknown, space?: number, maxDepth?: number) => string;
 /**
  * @description
  * Clone an JSON Object (any type) with max depth. \
  * This method goes through the object structure and replace children that goes deeper then the max Depth
- * @see http://vsc-base.org/#toString
+ * @see http://vsc-base.org/#maxDepthReplacer
  * @param obj
  * @param maxDepth
  * @param currentLevel