+++ vsc-base/out/vsc-base-raw.js
@@ -491,16 +491,16 @@ exports.getJSONCircularReplacer = () => {
  * @description
  * Stringify an object. \
  * Uses JSON.stringify and the circular ref safe replacer (see vsc.getJSONCircularReplacer)
- * @see http://vsc-base.org/#toString
+ * @see http://vsc-base.org/#toJSONString
  * @param obj
  * @param replacer
  * @param space
  * @vscType Raw
  * @debugTool Primary a debugging method.
- * @oneLineEx const objString = vsc.toString(someObject);
+ * @oneLineEx const objString = vsc.toJSONString(someObject);
  * @returns string
  */
-exports.toString = (obj, replacer = vsc.getJSONCircularReplacer(), space = 2, maxDepth = -1) => {
+exports.toJSONString = (obj, replacer = vsc.getJSONCircularReplacer(), space = 2, maxDepth = -1) => {
     if (maxDepth >= 0) {
         let maxDepthObj = exports.maxDepthReplacer(obj, maxDepth);
         return JSON.stringify(maxDepthObj, replacer, space);
@@ -511,7 +511,7 @@ exports.toString = (obj, replacer = vsc.getJSONCircularReplacer(), space = 2, ma
  * @description
  * Clone an JSON Object (any type) with max depth. \
  * This method goes through the object structure and replace children that goes deeper then the max Depth
- * @see http://vsc-base.org/#toString
+ * @see http://vsc-base.org/#maxDepthReplacer
  * @param obj
  * @param maxDepth
  * @param currentLevel